refactor: refactored posts layout page
